easy ts renames
